AJO - added auth-service file to clean up code
