+++ src/app/signup-modal/signup-modal.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { NgbModal, NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { Router } from '@angular/router';
+import { AuthService } from '../services/auth.service';
 import * as firebase from 'firebase';
 
 @Component({
@@ -9,37 +9,13 @@ import * as firebase from 'firebase';
   styleUrls: ['./signup-modal.component.css']
 })
 export class SignupModalComponent implements OnInit {
-	token: string;
 
-  constructor(public activeModal: NgbActiveModal, private router: Router) {}
+  constructor(public activeModal: NgbActiveModal, private authService: AuthService) {}
 
   ngOnInit() {
   }
 
   onSignup(name: string, email: string, username: string, password: string) {
-  	firebase.auth().createUserWithEmailAndPassword(email, password).catch(function(error) {
-  		let errCode = error.code;
-  		let errorMessage = error.message;
-  		alert(errCode + ' - ' + errorMessage);
-  	})
-  	.then(
-  		response => {	
-  			firebase.auth().currentUser.getIdToken()
-  			.then(
-  				(token: string) => this.token = token
-  			)
-  		})
-  	.catch(error => console.log(error))
-
-		let User = firebase.database().ref('users/');
-		let newUser = User.push()
-  	newUser.set({
-  		email: email,
-	    name: name,
-	    username: username,
-	    isAdmin: false
-	  })
-	  this.activeModal.close('Close click');
-	  this.router.navigate(['/home']);
-	}
+  	this.authService.newUser(name, email, username, password);
+  }
 }
